feat:前端使用vue初步重构，前后端分离(部分重构player页面)

+++ pokemon-portal/src/main/java/com/mimosa/pokemon/portal/entity/PageResponse.java
@@ -14,5 +14,5 @@ package com.mimosa.pokemon.portal.entity;
 
 import java.util.List;
 
-public record PageResponse<T>(int totalRecords, int page, int row, List<T> data) {
+public record PageResponse<T>(long totalRecords, int page, int row, List<T> data) {
 }
